backend/requirements.txt
# Airbot Backend Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
groq>=0.4.0
ollama>=0.1.0

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# RAG & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Excel Generation
openpyxl>=3.1.0

# Logging
loguru>=0.7.0

# Production Server
gunicorn>=21.0.0
